Make ToDebugString methods public and DebugViews easily obtainable (#20681)
Part of #20409
