+++ src/EFCore.Relational/Metadata/ColumnMappingExtensions.cs
@@ -5,30 +5,35 @@ using System.Text;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 
-namespace Microsoft.EntityFrameworkCore.Metadata.Internal
+namespace Microsoft.EntityFrameworkCore.Metadata
 {
     /// <summary>
-    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-    ///     any release. You should only use it directly in your code with extreme caution and knowing that
-    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+    ///     Extension methods for <see cref="IColumnMapping" />.
     /// </summary>
     public static class ColumnMappingExtensions
     {
         /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+        ///     <para>
+        ///         Creates a human-readable representation of the given metadata.
+        ///     </para>
+        ///     <para>
+        ///         Warning: Do not rely on the format of the returned string.
+        ///         It is designed for debugging only and may change arbitrarily between releases.
+        ///     </para>
         /// </summary>
+        /// <param name="columnMapping"> The metadata item. </param>
+        /// <param name="options"> Options for generating the string. </param>
+        /// <param name="indent"> The number of indent spaces to use before each new line. </param>
+        /// <returns> A human-readable representation. </returns>
         public static string ToDebugString(
             [NotNull] this IColumnMapping columnMapping,
             MetadataDebugStringOptions options,
-            [NotNull] string indent = "")
+            int indent = 0)
         {
             var builder = new StringBuilder();
+            var indentString = new string(' ', indent);
 
-            builder.Append(indent);
+            builder.Append(indentString);
 
             var singleLine = (options & MetadataDebugStringOptions.SingleLine) != 0;
             if (singleLine)
@@ -50,7 +55,7 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Internal
             if (!singleLine &&
                 (options & MetadataDebugStringOptions.IncludeAnnotations) != 0)
             {
-                builder.Append(columnMapping.AnnotationsToDebugString(indent + "  "));
+                builder.Append(columnMapping.AnnotationsToDebugString(indent + 2));
             }
 
             return builder.ToString();
