+++ src/EFCore.Relational/Metadata/DbFunctionExtensions.cs
@@ -6,31 +6,36 @@ using System.Text;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 
-namespace Microsoft.EntityFrameworkCore.Metadata.Internal
+namespace Microsoft.EntityFrameworkCore.Metadata
 {
     /// <summary>
-    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-    ///     any release. You should only use it directly in your code with extreme caution and knowing that
-    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+    ///     Extension methods for <see cref="IDbFunction" />.
     /// </summary>
     public static class DbFunctionExtensions
     {
         /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+        ///     <para>
+        ///         Creates a human-readable representation of the given metadata.
+        ///     </para>
+        ///     <para>
+        ///         Warning: Do not rely on the format of the returned string.
+        ///         It is designed for debugging only and may change arbitrarily between releases.
+        ///     </para>
         /// </summary>
+        /// <param name="function"> The metadata item. </param>
+        /// <param name="options"> Options for generating the string. </param>
+        /// <param name="indent"> The number of indent spaces to use before each new line. </param>
+        /// <returns> A human-readable representation. </returns>
         public static string ToDebugString(
             [NotNull] this IDbFunction function,
             MetadataDebugStringOptions options,
-            [NotNull] string indent = "")
+            int indent = 0)
         {
             var builder = new StringBuilder();
+            var indentString = new string(' ', indent);
 
             builder
-                .Append(indent)
+                .Append(indentString)
                 .Append("DbFunction: ");
 
             builder.Append(function.ReturnType.ShortDisplayName())
@@ -50,16 +55,16 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Internal
                 var parameters = function.Parameters.ToList();
                 if (parameters.Count != 0)
                 {
-                    builder.AppendLine().Append(indent).Append("  Parameters: ");
+                    builder.AppendLine().Append(indentString).Append("  Parameters: ");
                     foreach (var parameter in parameters)
                     {
-                        builder.AppendLine().Append(parameter.ToDebugString(options, indent + "    "));
+                        builder.AppendLine().Append(parameter.ToDebugString(options, indent + 4));
                     }
                 }
 
                 if ((options & MetadataDebugStringOptions.IncludeAnnotations) != 0)
                 {
-                    builder.Append(function.AnnotationsToDebugString(indent: indent + "  "));
+                    builder.Append(function.AnnotationsToDebugString(indent: indent + 2));
                 }
             }
