+++ src/EFCore.Relational/Metadata/TableExtensions.cs
@@ -6,31 +6,36 @@ using System.Text;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 
-namespace Microsoft.EntityFrameworkCore.Metadata.Internal
+namespace Microsoft.EntityFrameworkCore.Metadata
 {
     /// <summary>
-    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-    ///     any release. You should only use it directly in your code with extreme caution and knowing that
-    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+    ///     Extension methods for <see cref="ITable" />.
     /// </summary>
     public static class TableExtensions
     {
         /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+        ///     <para>
+        ///         Creates a human-readable representation of the given metadata.
+        ///     </para>
+        ///     <para>
+        ///         Warning: Do not rely on the format of the returned string.
+        ///         It is designed for debugging only and may change arbitrarily between releases.
+        ///     </para>
         /// </summary>
+        /// <param name="table"> The metadata item. </param>
+        /// <param name="options"> Options for generating the string. </param>
+        /// <param name="indent"> The number of indent spaces to use before each new line. </param>
+        /// <returns> A human-readable representation. </returns>
         public static string ToDebugString(
             [NotNull] this ITable table,
             MetadataDebugStringOptions options,
-            [NotNull] string indent = "")
+            int indent = 0)
         {
             var builder = new StringBuilder();
+            var indentString = new string(' ', indent);
 
             builder
-                .Append(indent)
+                .Append(indentString)
                 .Append("Table: ");
 
             if (table.Schema != null)
@@ -52,36 +57,36 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Internal
                 var mappings = table.EntityTypeMappings.ToList();
                 if (mappings.Count != 0)
                 {
-                    builder.AppendLine().Append(indent).Append("  EntityTypeMappings: ");
+                    builder.AppendLine().Append(indentString).Append("  EntityTypeMappings: ");
                     foreach (var mapping in mappings)
                     {
-                        builder.AppendLine().Append(mapping.ToDebugString(options, indent + "    "));
+                        builder.AppendLine().Append(mapping.ToDebugString(options, indent + 4));
                     }
                 }
 
                 var columns = table.Columns.ToList();
                 if (columns.Count != 0)
                 {
-                    builder.AppendLine().Append(indent).Append("  Columns: ");
+                    builder.AppendLine().Append(indentString).Append("  Columns: ");
                     foreach (var column in columns)
                     {
-                        builder.AppendLine().Append(column.ToDebugString(options, indent + "    "));
+                        builder.AppendLine().Append(column.ToDebugString(options, indent + 4));
                     }
                 }
 
                 var checkConstraints = table.CheckConstraints.ToList();
                 if (checkConstraints.Count != 0)
                 {
-                    builder.AppendLine().Append(indent).Append("  Check constraints: ");
+                    builder.AppendLine().Append(indentString).Append("  Check constraints: ");
                     foreach (var checkConstraint in checkConstraints)
                     {
-                        builder.AppendLine().Append(checkConstraint.ToDebugString(options, indent + "    "));
+                        builder.AppendLine().Append(checkConstraint.ToDebugString(options, indent + 4));
                     }
                 }
 
                 if ((options & MetadataDebugStringOptions.IncludeAnnotations) != 0)
                 {
-                    builder.Append(table.AnnotationsToDebugString(indent + "  "));
+                    builder.Append(table.AnnotationsToDebugString(indent + 2));
                 }
             }
