+++ src/EFCore.Relational/Metadata/ViewExtensions.cs
@@ -6,31 +6,36 @@ using System.Text;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 
-namespace Microsoft.EntityFrameworkCore.Metadata.Internal
+namespace Microsoft.EntityFrameworkCore.Metadata
 {
     /// <summary>
-    ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-    ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-    ///     any release. You should only use it directly in your code with extreme caution and knowing that
-    ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+    ///     Extension methods for <see cref="IView" />.
     /// </summary>
     public static class ViewExtensions
     {
         /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
+        ///     <para>
+        ///         Creates a human-readable representation of the given metadata.
+        ///     </para>
+        ///     <para>
+        ///         Warning: Do not rely on the format of the returned string.
+        ///         It is designed for debugging only and may change arbitrarily between releases.
+        ///     </para>
         /// </summary>
+        /// <param name="view"> The metadata item. </param>
+        /// <param name="options"> Options for generating the string. </param>
+        /// <param name="indent"> The number of indent spaces to use before each new line. </param>
+        /// <returns> A human-readable representation. </returns>
         public static string ToDebugString(
             [NotNull] this IView view,
             MetadataDebugStringOptions options,
-            [NotNull] string indent = "")
+            int indent = 0)
         {
             var builder = new StringBuilder();
+            var indentString = new string(' ', indent);
 
             builder
-                .Append(indent)
+                .Append(indentString)
                 .Append("View: ");
 
             if (view.Schema != null)
@@ -47,26 +52,26 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Internal
                 var mappings = view.EntityTypeMappings.ToList();
                 if (mappings.Count != 0)
                 {
-                    builder.AppendLine().Append(indent).Append("  EntityTypeMappings: ");
+                    builder.AppendLine().Append(indentString).Append("  EntityTypeMappings: ");
                     foreach (var mapping in mappings)
                     {
-                        builder.AppendLine().Append(mapping.ToDebugString(options, indent + "    "));
+                        builder.AppendLine().Append(mapping.ToDebugString(options, indent + 4));
                     }
                 }
 
                 var columns = view.Columns.ToList();
                 if (columns.Count != 0)
                 {
-                    builder.AppendLine().Append(indent).Append("  Properties: ");
+                    builder.AppendLine().Append(indentString).Append("  Properties: ");
                     foreach (var column in columns)
                     {
-                        builder.AppendLine().Append(column.ToDebugString(options, indent + "    "));
+                        builder.AppendLine().Append(column.ToDebugString(options, indent + 4));
                     }
                 }
 
                 if ((options & MetadataDebugStringOptions.IncludeAnnotations) != 0)
                 {
-                    builder.Append(view.AnnotationsToDebugString(indent: indent + "  "));
+                    builder.Append(view.AnnotationsToDebugString(indent + 2));
                 }
             }
