+++ src/EFCore/ChangeTracking/Internal/StateManager.cs
@@ -1254,16 +1254,5 @@ namespace Microsoft.EntityFrameworkCore.ChangeTracking.Internal
 
             @event?.Invoke(Context.ChangeTracker, new EntityStateChangedEventArgs(internalEntityEntry, oldState, newState));
         }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public virtual DebugView DebugView
-            => new DebugView(
-                () => this.ToDebugString(StateManagerDebugStringOptions.ShortDefault),
-                () => this.ToDebugString(StateManagerDebugStringOptions.LongDefault));
     }
 }
