+++ src/EFCore/Extensions/ForeignKeyExtensions.cs
@@ -2,9 +2,11 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using System.Text;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Microsoft.EntityFrameworkCore.Diagnostics;
+using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
@@ -78,5 +80,79 @@ namespace Microsoft.EntityFrameworkCore
         /// </returns>
         public static bool IsIntraHierarchical([NotNull] this IForeignKey foreignKey)
             => foreignKey.DeclaringEntityType.IsSameHierarchy(foreignKey.PrincipalEntityType);
+
+        /// <summary>
+        ///     <para>
+        ///         Creates a human-readable representation of the given metadata.
+        ///     </para>
+        ///     <para>
+        ///         Warning: Do not rely on the format of the returned string.
+        ///         It is designed for debugging only and may change arbitrarily between releases.
+        ///     </para>
+        /// </summary>
+        /// <param name="foreignKey"> The metadata item. </param>
+        /// <param name="options"> Options for generating the string. </param>
+        /// <param name="indent"> The number of indent spaces to use before each new line. </param>
+        /// <returns> A human-readable representation. </returns>
+        public static string ToDebugString(
+            [NotNull] this IForeignKey foreignKey,
+            MetadataDebugStringOptions options,
+            int indent = 0)
+        {
+            var builder = new StringBuilder();
+            var indentString = new string(' ', indent);
+
+            builder.Append(indentString);
+
+            var singleLine = (options & MetadataDebugStringOptions.SingleLine) != 0;
+            if (singleLine)
+            {
+                builder.Append("ForeignKey: ");
+            }
+
+            builder
+                .Append(foreignKey.DeclaringEntityType.DisplayName())
+                .Append(" ")
+                .Append(foreignKey.Properties.Format())
+                .Append(" -> ")
+                .Append(foreignKey.PrincipalEntityType.DisplayName())
+                .Append(" ")
+                .Append(foreignKey.PrincipalKey.Properties.Format());
+
+            if (foreignKey.IsUnique)
+            {
+                builder.Append(" Unique");
+            }
+
+            if (foreignKey.IsOwnership)
+            {
+                builder.Append(" Ownership");
+            }
+
+            if (foreignKey.PrincipalToDependent != null)
+            {
+                builder.Append(" ToDependent: ").Append(foreignKey.PrincipalToDependent.Name);
+            }
+
+            if (foreignKey.DependentToPrincipal != null)
+            {
+                builder.Append(" ToPrincipal: ").Append(foreignKey.DependentToPrincipal.Name);
+            }
+
+            if (foreignKey.DeleteBehavior != DeleteBehavior.NoAction)
+            {
+                builder
+                    .Append(" ")
+                    .Append(foreignKey.DeleteBehavior);
+            }
+
+            if (!singleLine &&
+                (options & MetadataDebugStringOptions.IncludeAnnotations) != 0)
+            {
+                builder.Append(foreignKey.AnnotationsToDebugString(indent + 2));
+            }
+
+            return builder.ToString();
+        }
     }
 }
