+++ src/EFCore/Extensions/IndexExtensions.cs
@@ -1,8 +1,11 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System.Linq;
+using System.Text;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
+using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
@@ -25,5 +28,56 @@ namespace Microsoft.EntityFrameworkCore
         /// <returns> The factory. </returns>
         public static IDependentKeyValueFactory<TKey> GetNullableValueFactory<TKey>([NotNull] this IIndex index)
             => index.AsIndex().GetNullableValueFactory<TKey>();
+
+        /// <summary>
+        ///     <para>
+        ///         Creates a human-readable representation of the given metadata.
+        ///     </para>
+        ///     <para>
+        ///         Warning: Do not rely on the format of the returned string.
+        ///         It is designed for debugging only and may change arbitrarily between releases.
+        ///     </para>
+        /// </summary>
+        /// <param name="index"> The metadata item. </param>
+        /// <param name="options"> Options for generating the string. </param>
+        /// <param name="indent"> The number of indent spaces to use before each new line. </param>
+        /// <returns> A human-readable representation. </returns>
+        public static string ToDebugString(
+            [NotNull] this IIndex index,
+            MetadataDebugStringOptions options,
+            int indent = 0)
+        {
+            var builder = new StringBuilder();
+            var indentString = new string(' ', indent);
+
+            builder.Append(indentString);
+
+            var singleLine = (options & MetadataDebugStringOptions.SingleLine) != 0;
+            if (singleLine)
+            {
+                builder.Append("Index: ");
+            }
+
+            builder
+                .AppendJoin(
+                    ", ",
+                    index.Properties.Select(
+                        p => singleLine
+                            ? p.DeclaringEntityType.DisplayName() + "." + p.Name
+                            : p.Name));
+
+            if (index.IsUnique)
+            {
+                builder.Append(" Unique");
+            }
+
+            if (!singleLine
+                && (options & MetadataDebugStringOptions.IncludeAnnotations) != 0)
+            {
+                builder.Append(index.AnnotationsToDebugString(indent + 2));
+            }
+
+            return builder.ToString();
+        }
     }
 }
