+++ src/EFCore/Extensions/KeyExtensions.cs
@@ -4,8 +4,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
+using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using Microsoft.EntityFrameworkCore.Utilities;
@@ -56,5 +58,55 @@ namespace Microsoft.EntityFrameworkCore
         /// <returns> <c>true</c> if the key is the primary key. </returns>
         public static bool IsPrimaryKey([NotNull] this IKey key)
             => key == key.DeclaringEntityType.FindPrimaryKey();
+
+        /// <summary>
+        ///     <para>
+        ///         Creates a human-readable representation of the given metadata.
+        ///     </para>
+        ///     <para>
+        ///         Warning: Do not rely on the format of the returned string.
+        ///         It is designed for debugging only and may change arbitrarily between releases.
+        ///     </para>
+        /// </summary>
+        /// <param name="key"> The metadata item. </param>
+        /// <param name="options"> Options for generating the string. </param>
+        /// <param name="indent"> The number of indent spaces to use before each new line. </param>
+        /// <returns> A human-readable representation. </returns>
+        public static string ToDebugString(
+            [NotNull] this IKey key,
+            MetadataDebugStringOptions options,
+            int indent = 0)
+        {
+            var builder = new StringBuilder();
+            var indentString = new string(' ', indent);
+
+            builder.Append(indentString);
+
+            var singleLine = (options & MetadataDebugStringOptions.SingleLine) != 0;
+            if (singleLine)
+            {
+                builder.Append("Key: ");
+            }
+
+            builder.AppendJoin(
+                ", ", key.Properties.Select(
+                    p => singleLine
+                        ? p.DeclaringEntityType.DisplayName() + "." + p.Name
+                        : p.Name));
+
+            if (key.IsPrimaryKey())
+            {
+                builder.Append(" PK");
+            }
+
+            if (!singleLine &&
+                (options & MetadataDebugStringOptions.IncludeAnnotations) != 0)
+            {
+                builder.Append(key.AnnotationsToDebugString(indent + 2));
+            }
+
+            return builder.ToString();
+        }
+
     }
 }
