+++ src/EFCore/Infrastructure/AnnotatableExtensions.cs
@@ -39,9 +39,9 @@ namespace Microsoft.EntityFrameworkCore.Infrastructure
         ///     Gets the debug string for all annotations declared on the object.
         /// </summary>
         /// <param name="annotatable"> The object to get the annotations to print in debug string. </param>
-        /// <param name="indent"> The indentation to use when printing. </param>
+        /// <param name="indent"> The number of indent spaces to use before each new line. </param>
         /// <returns> Debug string representation of all annotations. </returns>
-        public static string AnnotationsToDebugString([NotNull] this IAnnotatable annotatable, [NotNull] string indent = "")
+        public static string AnnotationsToDebugString([NotNull] this IAnnotatable annotatable, int indent = 0)
         {
             var annotations = annotatable.GetAnnotations().ToList();
             if (annotations.Count == 0)
@@ -50,13 +50,14 @@ namespace Microsoft.EntityFrameworkCore.Infrastructure
             }
 
             var builder = new StringBuilder();
+            var indentString = new string(' ', indent);
 
-            builder.AppendLine().Append(indent).Append("Annotations: ");
+            builder.AppendLine().Append(indentString).Append("Annotations: ");
             foreach (var annotation in annotations)
             {
                 builder
                     .AppendLine()
-                    .Append(indent)
+                    .Append(indentString)
                     .Append("  ")
                     .Append(annotation.Name)
                     .Append(": ")
