+++ src/EFCore/Metadata/Internal/KeyExtensions.cs
@@ -2,12 +2,9 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Text;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.ChangeTracking.Internal;
-using Microsoft.EntityFrameworkCore.Infrastructure;
 
 namespace Microsoft.EntityFrameworkCore.Metadata.Internal
 {
@@ -42,47 +39,6 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Internal
             }
 
             return index == key.Properties.Count ? -1 : index;
-        }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public static string ToDebugString(
-            [NotNull] this IKey key,
-            MetadataDebugStringOptions options,
-            [NotNull] string indent = "")
-        {
-            var builder = new StringBuilder();
-
-            builder.Append(indent);
-
-            var singleLine = (options & MetadataDebugStringOptions.SingleLine) != 0;
-            if (singleLine)
-            {
-                builder.Append("Key: ");
-            }
-
-            builder.AppendJoin(
-                ", ", key.Properties.Select(
-                    p => singleLine
-                        ? p.DeclaringEntityType.DisplayName() + "." + p.Name
-                        : p.Name));
-
-            if (key.IsPrimaryKey())
-            {
-                builder.Append(" PK");
-            }
-
-            if (!singleLine &&
-                (options & MetadataDebugStringOptions.IncludeAnnotations) != 0)
-            {
-                builder.Append(key.AnnotationsToDebugString(indent + "  "));
-            }
-
-            return builder.ToString();
         }
 
         /// <summary>
