+++ src/EFCore/Metadata/Internal/ModelExtensions.cs
@@ -4,9 +4,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Text;
 using JetBrains.Annotations;
-using Microsoft.EntityFrameworkCore.Infrastructure;
 
 namespace Microsoft.EntityFrameworkCore.Metadata.Internal
 {
@@ -35,44 +33,6 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Internal
         /// </summary>
         public static IEnumerable<IEntityType> GetRootEntityTypes([NotNull] this IModel model)
             => model.GetEntityTypes().Where(e => e.BaseType == null);
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public static string ToDebugString(
-            [NotNull] this IModel model,
-            MetadataDebugStringOptions options,
-            [NotNull] string indent = "")
-        {
-            var builder = new StringBuilder();
-
-            builder.Append(indent).Append("Model: ");
-
-            if (model.GetPropertyAccessMode() != PropertyAccessMode.PreferField)
-            {
-                builder.Append(" PropertyAccessMode.").Append(model.GetPropertyAccessMode());
-            }
-
-            if (model.GetChangeTrackingStrategy() != ChangeTrackingStrategy.Snapshot)
-            {
-                builder.Append(" ChangeTrackingStrategy.").Append(model.GetChangeTrackingStrategy());
-            }
-
-            foreach (var entityType in model.GetEntityTypes())
-            {
-                builder.AppendLine().Append(entityType.ToDebugString(options, indent + "  "));
-            }
-
-            if ((options & MetadataDebugStringOptions.IncludeAnnotations) != 0)
-            {
-                builder.Append(model.AnnotationsToDebugString(indent));
-            }
-
-            return builder.ToString();
-        }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
