+++ src/EFCore/Metadata/Internal/PropertyExtensions.cs
@@ -1,13 +1,10 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Text;
 using JetBrains.Annotations;
-using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.ValueGeneration;
 
 namespace Microsoft.EntityFrameworkCore.Metadata.Internal
@@ -143,129 +140,6 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Internal
                 || property.IsKey()
                 || property.IsForeignKey()
                 || property.IsUniqueIndex();
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public static string ToDebugString(
-            [NotNull] this IProperty property,
-            MetadataDebugStringOptions options,
-            [NotNull] string indent = "")
-        {
-            var builder = new StringBuilder();
-
-            builder.Append(indent);
-
-            var singleLine = (options & MetadataDebugStringOptions.SingleLine) != 0;
-            if (singleLine)
-            {
-                builder.Append($"Property: {property.DeclaringEntityType.DisplayName()}.");
-            }
-
-            builder.Append(property.Name).Append(" (");
-
-            var field = property.GetFieldName();
-            if (field == null)
-            {
-                builder.Append("no field, ");
-            }
-            else if (!field.EndsWith(">k__BackingField", StringComparison.Ordinal))
-            {
-                builder.Append(field).Append(", ");
-            }
-
-            builder.Append(property.ClrType.ShortDisplayName()).Append(")");
-
-            if (property.IsShadowProperty())
-            {
-                builder.Append(" Shadow");
-            }
-
-            if (!property.IsNullable)
-            {
-                builder.Append(" Required");
-            }
-
-            if (property.IsPrimaryKey())
-            {
-                builder.Append(" PK");
-            }
-
-            if (property.IsForeignKey())
-            {
-                builder.Append(" FK");
-            }
-
-            if (property.IsKey()
-                && !property.IsPrimaryKey())
-            {
-                builder.Append(" AlternateKey");
-            }
-
-            if (property.IsIndex())
-            {
-                builder.Append(" Index");
-            }
-
-            if (property.IsConcurrencyToken)
-            {
-                builder.Append(" Concurrency");
-            }
-
-            if (property.GetBeforeSaveBehavior() != PropertySaveBehavior.Save)
-            {
-                builder.Append(" BeforeSave:").Append(property.GetBeforeSaveBehavior());
-            }
-
-            if (property.GetAfterSaveBehavior() != PropertySaveBehavior.Save)
-            {
-                builder.Append(" AfterSave:").Append(property.GetAfterSaveBehavior());
-            }
-
-            if (property.ValueGenerated != ValueGenerated.Never)
-            {
-                builder.Append(" ValueGenerated.").Append(property.ValueGenerated);
-            }
-
-            if (property.GetMaxLength() != null)
-            {
-                builder.Append(" MaxLength(").Append(property.GetMaxLength()).Append(")");
-            }
-
-            if (property.IsUnicode() == false)
-            {
-                builder.Append(" Ansi");
-            }
-
-            if (property.GetPropertyAccessMode() != PropertyAccessMode.PreferField)
-            {
-                builder.Append(" PropertyAccessMode.").Append(property.GetPropertyAccessMode());
-            }
-
-            if ((options & MetadataDebugStringOptions.IncludePropertyIndexes) != 0)
-            {
-                var indexes = property.GetPropertyIndexes();
-                if (indexes != null)
-                {
-                    builder.Append(" ").Append(indexes.Index);
-                    builder.Append(" ").Append(indexes.OriginalValueIndex);
-                    builder.Append(" ").Append(indexes.RelationshipIndex);
-                    builder.Append(" ").Append(indexes.ShadowIndex);
-                    builder.Append(" ").Append(indexes.StoreGenerationIndex);
-                }
-            }
-
-            if (!singleLine &&
-                (options & MetadataDebugStringOptions.IncludeAnnotations) != 0)
-            {
-                builder.Append(property.AnnotationsToDebugString(indent + "  "));
-            }
-
-            return builder.ToString();
-        }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
