+++ src/EFCore/Metadata/Internal/ServicePropertyExtensions.cs
@@ -2,9 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.Runtime.CompilerServices;
-using System.Text;
 using JetBrains.Annotations;
-using Microsoft.EntityFrameworkCore.Infrastructure;
 
 namespace Microsoft.EntityFrameworkCore.Metadata.Internal
 {
@@ -24,49 +22,6 @@ namespace Microsoft.EntityFrameworkCore.Metadata.Internal
         /// </summary>
         public static ServiceParameterBinding GetParameterBinding([NotNull] this IServiceProperty serviceProperty)
             => serviceProperty.AsServiceProperty().ParameterBinding;
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        public static string ToDebugString(
-            [NotNull] this IServiceProperty serviceProperty,
-            MetadataDebugStringOptions options,
-            [NotNull] string indent = "")
-        {
-            var builder = new StringBuilder();
-
-            builder.Append(indent);
-
-            var singleLine = (options & MetadataDebugStringOptions.SingleLine) != 0;
-            if (singleLine)
-            {
-                builder.Append("Service property: ").Append(serviceProperty.DeclaringType.DisplayName()).Append(".");
-            }
-
-            builder.Append(serviceProperty.Name);
-
-            if (serviceProperty.GetFieldName() == null)
-            {
-                builder.Append(" (no field, ");
-            }
-            else
-            {
-                builder.Append(" (").Append(serviceProperty.GetFieldName()).Append(", ");
-            }
-
-            builder.Append(serviceProperty.ClrType?.ShortDisplayName()).Append(")");
-
-            if (!singleLine &&
-                (options & MetadataDebugStringOptions.IncludeAnnotations) != 0)
-            {
-                builder.Append(serviceProperty.AnnotationsToDebugString(indent + "  "));
-            }
-
-            return builder.ToString();
-        }
 
         /// <summary>
         ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
